repository: prepare StreamPack refactor
